Modify ApiCrypto class and create_target_table.sql

+++ dags/sql/create_target_tables.sql
@@ -4,10 +4,10 @@ CREATE TABLE IF NOT EXISTS crypto
 	time_period_end date,
 	time_open date,
 	time_close date,
-	price_open NUMERIC(10,5),
-	price_high NUMERIC(10,5),
-	price_low NUMERIC(10,5),
-	price_close NUMERIC(10,5),
-	volume_traded NUMERIC(10,5),
+	price_open DOUBLE PRECISION,
+	price_high DOUBLE PRECISION,
+	price_low DOUBLE PRECISION,
+	price_close DOUBLE PRECISION,
+	volume_traded DOUBLE PRECISION,
 	trades_count integer
 )
